rename package data to entities move JsonLoad to package repository
