+++ app/src/main/java/ru/aevd/androidacademymovieapp/TransactionsFragmentClicks.kt
@@ -1,6 +1,6 @@
 package ru.aevd.androidacademymovieapp
 
-import ru.aevd.androidacademymovieapp.data.Movie
+import ru.aevd.androidacademymovieapp.entities.Movie
 
 interface TransactionsFragmentClicks {
     fun showMovieDetails(movie: Movie)
