+++ app/src/main/java/ru/aevd/androidacademymovieapp/entities/Genre.kt
@@ -1,0 +1,10 @@
+package ru.aevd.androidacademymovieapp.entities
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Genre(
+    val id: Int,
+    val name: String
+) : Parcelable
