+++ app/src/main/java/ru/aevd/androidacademymovieapp/entities/Movie.kt
@@ -1,4 +1,4 @@
-package ru.aevd.androidacademymovieapp.data
+package ru.aevd.androidacademymovieapp.entities
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
