+++ app/src/main/java/ru/aevd/androidacademymovieapp/repository/MoviePersistent.kt
@@ -1,8 +1,7 @@
 package ru.aevd.androidacademymovieapp.repository
 
 import android.content.Context
-import ru.aevd.androidacademymovieapp.data.Movie
-import ru.aevd.androidacademymovieapp.data.loadMovies
+import ru.aevd.androidacademymovieapp.entities.Movie
 
 class MoviePersistent(private val context: Context) {
     suspend fun loadMovies(): List<Movie> = loadMovies(context)
