+++ app/src/main/java/ru/aevd/androidacademymovieapp/viewmodels/MoviesDetailsViewModelFactory.kt
@@ -2,7 +2,7 @@ package ru.aevd.androidacademymovieapp.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import ru.aevd.androidacademymovieapp.data.Movie
+import ru.aevd.androidacademymovieapp.entities.Movie
 
 class MoviesDetailsViewModelFactory(
         private val movie: Movie
