+++ app/src/main/java/ru/aevd/androidacademymovieapp/viewmodels/MoviesListViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
 import ru.aevd.androidacademymovieapp.repository.GetMoviesUseCase
-import ru.aevd.androidacademymovieapp.data.Movie
+import ru.aevd.androidacademymovieapp.entities.Movie
 
 @Suppress("unused")
 class MoviesListViewModel(
